+++ include/cwru_davinci_grasp/davinci_needle_pose_publisher.h
@@ -51,7 +51,8 @@ public :
 
   DummyNeedleModifier
   (
-  const ros::NodeHandle &nh
+  const ros::NodeHandle &nh,
+  const ros::NodeHandle &nhPriv
   );
 
   virtual ~DummyNeedleModifier(){}
@@ -94,6 +95,7 @@ protected:
 
 protected:
   ros::NodeHandle               m_NodeHandle;
+  ros::NodeHandle               m_PrivNodeHandle;
   ros::ServiceClient            m_NeedlePoseClient;
   ros::ServiceClient            m_NeedlePoseModifier;
 
@@ -106,196 +108,13 @@ protected:
   std::random_device            m_RandSeed;
 };
 
-DummyNeedleModifier::DummyNeedleModifier
-(
-const ros::NodeHandle &nh
-) 
-: m_NodeHandle(nh)
-{
-  initialize();
-}
-
-void DummyNeedleModifier::initialize
-(
-)
-{
-  m_NeedlePoseClient = m_NodeHandle.serviceClient<gazebo_msgs::GetModelState>("/gazebo/get_model_state");
-  m_NeedlePoseModifier = m_NodeHandle.serviceClient<gazebo_msgs::SetModelState>("/gazebo/set_model_state");
-
-  m_NeedleState.request.model_name = "needle_r";
-  m_NeedleState.request.relative_entity_name = "davinci_endo_cam_l";
-
-  m_PerturbedNeedleState.request.model_state.model_name = "needle_r";
-  m_PerturbedNeedleState.request.model_state.reference_frame = "davinci_endo_cam_l";
-}
-
-bool DummyNeedleModifier::getCurrentNeedlePose
-(
-Eigen::Affine3d& currentNeedlePose
-)
-{
-  if (!m_NeedlePoseClient.call(m_NeedleState))
-  {
-    ROS_WARN("DummyNeedleModifier: Failed getting neeedle pose from gazebo");
-    return false;
-  }
-  tf::poseMsgToEigen(m_NeedleState.response.pose, currentNeedlePose);
-
-  return true;
-}
-
-bool DummyNeedleModifier::perturbNeedlePose
-(
-double perturbRadian,
-const cwru_davinci_grasp::GraspInfo& selectedGrasp,
-const Eigen::Affine3d& idealNeedlePose,
-bool useIdealNdlPose,
-char rotAxis,
-bool randomRotAxis
-)
-{
-  Eigen::Affine3d needlePose;
-  if (useIdealNdlPose)
-  {
-    needlePose = idealNeedlePose;
-  }
-  else
-  {
-    if (!m_NeedlePoseClient.call(m_NeedleState))
-    {
-      ROS_WARN("DummyNeedleModifier: Failed getting neeedle pose from gazebo");
-      return false;
-    }
-    tf::poseMsgToEigen(m_NeedleState.response.pose, needlePose);
-  }
-
-  double needle_radius = 0.0125;  // TODO
-  double radius = selectedGrasp.graspParamInfo.param_3;
-  double x = needle_radius * cos(radius);
-  double y = needle_radius * sin(radius);
-
-  // Transform vector into world frame
-  Eigen::Vector3d vecFromNeedleOriginToGraspPointWrtWorldFrame = needlePose.linear() * Eigen::Vector3d(x, y, 0.0);
-  vecFromNeedleOriginToGraspPointWrtWorldFrame.normalize();
-
-  // Transform point into world frame
-  Eigen::Vector3d graspPointLocationWrtWorldFrame = needlePose * Eigen::Vector3d(x, y, 0.0);
-
-  Eigen::Vector3d needleFrameZaxisProjectedOnWorldFrame = needlePose.rotation().col(2);
-
-  Eigen::Vector3d perturbAxisOfRotWrtWorldFrame = needleFrameZaxisProjectedOnWorldFrame;
-
-  if (!randomRotAxis)
-  {
-    switch (rotAxis)
-    {
-      case 'z' :
-        break;
-      case 't' :
-        perturbAxisOfRotWrtWorldFrame = vecFromNeedleOriginToGraspPointWrtWorldFrame.cross(needleFrameZaxisProjectedOnWorldFrame);
-        ROS_INFO("DummyNeedleModifier: Perturbation is around Tangential Axis");
-        // perturbRadian -= 0.08;  // compensation, this is a number from statistics
-        break;
-      case 'r' :
-        perturbAxisOfRotWrtWorldFrame = vecFromNeedleOriginToGraspPointWrtWorldFrame;
-        ROS_INFO("DummyNeedleModifier: Perturbation is around Radial Axis");
-        // perturbRadian -= 0.03;  // compensation, this is a number from statistics
-        break;
-    }
-  }
-  else
-  {
-    if (m_Distribution(m_RandSeed))
-    {
-      if(m_Distribution(m_RandSeed))
-      {
-        rotAxis = 't';
-        perturbAxisOfRotWrtWorldFrame = vecFromNeedleOriginToGraspPointWrtWorldFrame.cross(needleFrameZaxisProjectedOnWorldFrame);
-        ROS_INFO("DummyNeedleModifier: Perturbation is around Tangential Axis");
-        // perturbRadian -= 0.08;  // compensation, this is a number from statistics
-      }
-      else
-      {
-        rotAxis = 'r';
-        perturbAxisOfRotWrtWorldFrame = vecFromNeedleOriginToGraspPointWrtWorldFrame;
-        ROS_INFO("DummyNeedleModifier: Perturbation is around Radial Axis");
-        // perturbRadian -= 0.03;  // compensation, this is a number from statistics
-      }
-    }
-  }
-
-  if (rotAxis == 'z')
-    ROS_INFO("DummyNeedleModifier: Perturbation is around Vertical Axis");
-
-  Eigen::AngleAxisd perturbRotationMatWrtWorldFrame(perturbRadian, perturbAxisOfRotWrtWorldFrame);
-
-  Eigen::Vector3d needleLocationWrtWorldFrame = needlePose.translation();
-  // Eigen::Vector3d graspPointLocationWrtWorldFrame = needleLocationWrtWorldFrame + vecFromNeedleOriginToGraspPointWrtWorldFrame;
-
-  Eigen::Affine3d TM1(Eigen::Affine3d::Identity());     TM1.translation() = graspPointLocationWrtWorldFrame;
-
-  Eigen::Affine3d TM2(perturbRotationMatWrtWorldFrame);
-
-  Eigen::Affine3d TM3(Eigen::Affine3d::Identity());     TM3.translation() = -graspPointLocationWrtWorldFrame;
-
-  Eigen::Affine3d perturbedNeedlePose = TM1 * TM2 * TM3 * needlePose;
-
-  tf::poseEigenToMsg(perturbedNeedlePose, m_PerturbedNeedleState.request.model_state.pose);
-  if (!m_NeedlePoseModifier.call(m_PerturbedNeedleState))
-  {
-    ROS_WARN("DummyNeedleModifier: Failed setting perturbed neeedle pose to gazebo");
-    return false;
-  }
-
-  return true;
-}
-
-bool DummyNeedleModifier::setNeedlePose
-(
-double x,
-double y,
-double z,
-double qw,
-double qx,
-double qy,
-double qz
-)
-{
-  m_PerturbedNeedleState.request.model_state.pose.position.x = x;
-  m_PerturbedNeedleState.request.model_state.pose.position.y = y;
-  m_PerturbedNeedleState.request.model_state.pose.position.z = z;
-
-  m_PerturbedNeedleState.request.model_state.pose.orientation.w = qw;
-  m_PerturbedNeedleState.request.model_state.pose.orientation.x = qx;
-  m_PerturbedNeedleState.request.model_state.pose.orientation.y = qy;
-  m_PerturbedNeedleState.request.model_state.pose.orientation.z = qz;
-  if (!m_NeedlePoseModifier.call(m_PerturbedNeedleState))
-  {
-    ROS_WARN("DummyNeedleModifier: Failed to set needle pose");
-    return false;
-  }
-
-  return true;
-}
-
-void DummyNeedleModifier::radianOfChange
-(
-double& radOfChange,
-const Eigen::Affine3d& t1,
-const Eigen::Affine3d& t2
-)
-{
-  // q_w_ri, q_w_rp, q_ri_rp = q_ri_w * q_w_rp = transpose(q_w_ri) * q_w_rp;
-  const Eigen::Matrix3d rot = t1.linear().inverse() * t2.linear();
-  radOfChange = acos( (rot.trace() - 1) / 2 );
-}
-
 class DummyNeedleTracker : public DummyNeedleModifier
 {
 public :
   DummyNeedleTracker
   (
-  const ros::NodeHandle &nh
+  const ros::NodeHandle &nh,
+  const ros::NodeHandle &nhPriv
   );
 
   virtual ~DummyNeedleTracker(){}
@@ -314,74 +133,21 @@ protected:
   Eigen::Affine3d                       m_IdealNeedlePose;
 };
 
-DummyNeedleTracker::DummyNeedleTracker
-(
-const ros::NodeHandle &nh
-)
- : DummyNeedleModifier(nh)
-{
-  m_NeedlePosePub = m_NodeHandle.advertise<geometry_msgs::PoseStamped>("/updated_needle_pose", 1000);
-}
-
-bool DummyNeedleTracker::publishNeedlePose()
-{
-  if (!getNeedlePose())
-    return false;
-
-  m_NeedlePosePub.publish(m_StampedNeedlePose);
-  return true;
-}
-
-bool DummyNeedleTracker::getNeedlePose
-(
-)
-{
-  if (!m_NeedlePoseClient.call(m_NeedleState))
-  {
-    ROS_WARN("DummyNeedleModifier: Failed getting neeedle pose from gazebo");
-    return false;
-  }
-
-  m_StampedNeedlePose.header.frame_id = m_NeedleState.response.header.frame_id;
-  m_StampedNeedlePose.header.stamp = ros::Time::now();
-  m_StampedNeedlePose.pose = m_NeedleState.response.pose;
-  return true;
-}
-
 class DummyNeedleTrackerWithDepthNoise : public DummyNeedleTracker
 {
 public :
   DummyNeedleTrackerWithDepthNoise
   (
-  const ros::NodeHandle &nh
-  ) : DummyNeedleTracker(nh)
-  {
-    //  blank
-  }
+  const ros::NodeHandle &nh,
+  const ros::NodeHandle &nhPriv
+  );
 
   virtual ~DummyNeedleTrackerWithDepthNoise(){}
 
-  virtual bool publishNeedlePose() override
-  {
-    if (!getNeedlePose())
-      return false;
-
-    tf::poseMsgToEigen(m_StampedNeedlePose.pose, m_IdealNeedlePose);
-
-    m_NoiseVec.z() = m_TransNoise(m_RandSeed);
-
-    // (I, T.z + std::normal_distribution(0.002, 0.0001)) * G_wn
-    m_NeedlePoseWithNoise = m_IdealNeedlePose.pretranslate(m_NoiseVec);
-
-    tf::poseEigenToMsg(m_NeedlePoseWithNoise, m_StampedNeedlePose.pose);
-
-    m_NeedlePosePub.publish(m_StampedNeedlePose);
-
-    return true;
-  }
+  virtual bool publishNeedlePose() override;
 
 protected:
-  std::normal_distribution<double>      m_TransNoise{0.002, 0.0001};
+  std::normal_distribution<double>      m_TransNoise;
 
   Eigen::Vector3d                       m_NoiseVec = Eigen::Vector3d::Zero();
 };
@@ -391,33 +157,16 @@ class DummyNeedleTrackerWithRotationNoise : public DummyNeedleTracker
 public :
   DummyNeedleTrackerWithRotationNoise
   (
-  const ros::NodeHandle &nh
-  ) : DummyNeedleTracker(nh)
-  {
-    //  blank
-  }
+  const ros::NodeHandle &nh,
+  const ros::NodeHandle &nhPriv
+  );
 
   virtual ~DummyNeedleTrackerWithRotationNoise(){}
 
-  virtual bool publishNeedlePose() override
-  {
-    if (!getNeedlePose())
-      return false;
-
-    tf::poseMsgToEigen(m_StampedNeedlePose.pose, m_IdealNeedlePose);
-
-    // G_wn * (Rot_o/n, T == 0)
-    // (I, T.z + 0.002) * G_wn *(Rot_o/n, T == 0)
-    m_NoiseMat = Eigen::AngleAxisd(m_OrientNoise(m_RandSeed), m_Zaxis);
-    m_NeedlePoseWithNoise = m_IdealNeedlePose.rotate(m_NoiseMat);
-
-    tf::poseEigenToMsg(m_NeedlePoseWithNoise, m_StampedNeedlePose.pose);
-
-    m_NeedlePosePub.publish(m_StampedNeedlePose);
-  }
+  virtual bool publishNeedlePose() override;
 
 protected:
-  std::normal_distribution<double>      m_OrientNoise{0.02, 0.001};
+  std::normal_distribution<double>      m_OrientNoise;
 
   Eigen::Vector3d                       m_Zaxis = Eigen::Vector3d::UnitZ();
   Eigen::Matrix3d                       m_NoiseMat;
