+++ launch/davinci_needle_pose_publisher.launch
@@ -1,0 +1,11 @@
+<launch>
+    # specify the needle tracking mode
+    <arg name = "pub_mode" default = "0"/>
+
+    <node name="davinci_needle_pose_publisher" pkg="cwru_davinci_grasp" type="davinci_needle_pose_publisher_main" args="$(arg pub_mode)" respawn="false" output="screen">
+        <param name="trans_noise_mean"    type="double" value="0.002"/>
+        <param name="trans_noise_stddev"  type="double" value="0.0001"/>
+        <param name="orient_noise_mean"   type="double" value="0.02"/>
+        <param name="orient_noise_stddev" type="double" value="0.001"/>
+    </node>
+</launch>
